zephyr-build: devicetree: Move augments into module
The hard-coded augments in augment.rs are no long used, so remove them
all, and move the config version into this file to avoid needing a
separate module, just for the trait.
