+++ zephyr-build/src/devicetree.rs
@@ -24,12 +24,11 @@ use ordmap::OrdMap;
 use std::{cell::RefCell, collections::BTreeMap, path::Path, rc::Rc};
 
 mod augment;
-pub mod config;
 mod ordmap;
 mod output;
 mod parse;
 
-pub use augment::Augment;
+pub use augment::{Augment, load_augments};
 
 pub struct DeviceTree {
     /// The root of the tree.
